+++ apps/openci_worker_cli/lib/constants.dart
@@ -1,4 +1,4 @@
-const version = '0.10.12';
+const version = '0.10.13';
 
 // macOS (Lume VM)
 const baseVmName = 'tahoe-base_v1.1.1';